==> globalprep/prs_land-based_nutrient/v2022/STEP2_manure/STEP4_prop_manure_applied_pasture.Rmd <==
---
title: "OHI2022: STEP4: create rasters for prop N going to soils"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '../../../workflow/templates/ohi_hdr.html'
pdf_document:
  toc: true
editor_options: 
  chunk_output_type: console
---

## Summary
Creates rasters for the proportion of initial manure N that is applied to soils and also that is left on pastures

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(here)
library(countrycode)
library(stringr)
library(doParallel)
library(readr)
library(tidyverse)

source(here('workflow/R/common.R'))

region_data()

nutrient_d2021_anx <- file.path("/home/shares/ohi/git-annex/globalprep/_raw_data/FAOSTAT/manure_nutrient/d2021/")
git_manure <- file.path(here(), "globalprep/prs_land-based_nutrient/v2021/STEP2_manure")
git_crop <- file.path(here(), "globalprep/prs_land-based_nutrient/v2021/STEP1_crop")
location_tifs_path <- file.path("/home/shares/ohi/git-annex/globalprep/_raw_data/halpern_food_systems/d2021/animal_farm/location_tifs")
prep <- file.path("/home/shares/ohi/git-annex/globalprep/prs_land-based_nutrient/v2021")

source(here(git_crop, "R/common_spatial.R")) # read in necessary spatial files 


un <- read_csv(here("globalprep/spatial/v2021/UNSD_Methodology.csv")) %>% 
  dplyr::select(iso3c, Global_Name, Region_Name, Sub_region_Name, Intermediate_Region_Name)

un_dev <- read_csv(here("globalprep/spatial/v2021/UNSD_Methodology.csv")) %>% 
  right_join(food_rgns_xy) %>% 
  dplyr::select(x,y, iso3c, Developed_Developing)

gleam_tables <- read_csv(here(git_manure, "raw/compiled_gleam_tables.csv")) %>%
  filter(!mms_specific %in% c("Burned for fuel", "Pasture/Range/Paddock")) %>% 
  mutate(mms_specific = str_trim(mms_specific, "both"))

ippc_table <- read_csv(here::here(git_manure, "raw/IPCC_table_10_ch10_vol4.csv")) %>% 
  mutate(mms_specific = str_trim(mms_specific, "both"))

exc_df_list <- list.files(file.path(prep, "animal_farm_int/N_from_excretion_manure"),  full = TRUE)

```

Loop through the N excretion rasters 
```{r}
exc_df_distinct <- as.data.frame(exc_df_list) %>%
  rename("filepath" = "exc_df_list") %>%
  mutate(rast_file = gsub('.*/', '', filepath)) %>%
  separate(col = rast_file, sep = "_", into = c("animal_name", "system_name", "product_name", "nutrient", "process", "year")) %>%
  dplyr::select(-nutrient, -process, -year, -filepath) %>%
  distinct() 


registerDoParallel(10)   
foreach(file = 1:nrow(exc_df_distinct)) %dopar% {

  ############  
  ## Get the mapped head counts for each system.
  ## (only really used in step 1, to check data)
  
     df = exc_df_distinct[file, ]
     
     # df = exc_df_distinct[5, ]
  
    ## grab the names of all the parts from the file
    animal_name <- df$animal_name
    system_name <- df$system_name
    product_name <- df$product_name
    
    ## This describes the proportion of waste going to each manure fate based on 
    ## country and animal system and the fraction that leaches
    mms_allocate <- gleam_tables %>% 
      filter(mms_specific !="Daily spread") %>% 
      filter(animal == animal_name) %>% 
      filter(product == product_name) %>% 
      filter(production_system == system_name) %>% 
      dplyr::select(-ef_direct, -ef_dir_unit, -frac_gas, -percentage_leach) %>%
      unique()
    
   #length(unique(mms_allocate$iso3c)) ## missing some countries so will need to gapfill. probably later down the line when we have summarized by total N lost for each country (not mms) and then calculate a rate instead of gapfilling percentages
   
    N_applied_post_mms_frac <- left_join(mms_allocate, ippc_table, by = c("mms_specific", "animal", "product")) %>% 
      mutate(frac_applied_post_mms = (to_mms_percentage/100)*(1-frac_N_lost)) %>% 
      group_by(iso3c, country, animal, production_system, product) %>% 
      dplyr::summarise(frac_applied_post_mms = sum(frac_applied_post_mms)) %>% 
      ungroup() %>% 
      right_join(food_rgns_xy) %>% 
      dplyr::select(x, y, iso3c, animal, product, system = production_system, frac_applied_post_mms )
      # dplyr::select(x,y, frac_applied_post_mms) %>%
      # rasterFromXYZ(crs = food_crs)
      # 
      # plot(N_applied_post_mms_frac)
    
      mms_spread <- gleam_tables %>% 
      filter(mms_specific == "Daily spread") %>% 
      filter(animal == animal_name) %>% 
      filter(product == product_name) %>% 
      filter(production_system == system_name) %>% 
      mutate(to_mms_percentage = to_mms_percentage/100) %>% 
      dplyr::select(iso3c, Country = country, animal, system = production_system, product, to_mms_percentage) %>% 
      unique() ## IND and PSE are duplicated.... 
    
    N_direct <- left_join(food_rgns_xy, mms_spread)  %>% 
      rename(N_direct_applied = to_mms_percentage)  %>% 
      dplyr::select(x, y, iso3c, animal, product, system, N_direct_applied)# %>%
      # dplyr::select(x,y, N_direct_applied) %>%
      # rasterFromXYZ(crs = food_crs)
    #plot(N_direct)
      
    # N_left_rast <- sum(N_applied_post_mms_frac, N_direct, na.rm = TRUE)
    N_left <- left_join(N_applied_post_mms_frac, N_direct) %>% 
      mutate(prop_N_applied = frac_applied_post_mms + N_direct_applied) %>% 
      dplyr::select(-frac_applied_post_mms, -N_direct_applied) #%>%
    #   dplyr::select(x,y, prop_N_applied) %>%
    #   rasterFromXYZ(crs=food_crs)
    # plot(N_left)
    
    N_left_gf <- left_join(N_left, un, by = "iso3c") %>% 
      group_by(Intermediate_Region_Name) %>% 
      mutate(prop_N_applied = ifelse(is.na(prop_N_applied), mean(prop_N_applied, na.rm = TRUE), prop_N_applied)) %>% 
      ungroup()%>% 
      group_by(Sub_region_Name) %>% 
      mutate(prop_N_applied = ifelse(is.na(prop_N_applied), mean(prop_N_applied, na.rm = TRUE), prop_N_applied)) %>% 
      ungroup() %>% 
      group_by(Region_Name) %>% 
      mutate(prop_N_applied = ifelse(is.na(prop_N_applied), mean(prop_N_applied, na.rm = TRUE), prop_N_applied)) %>% 
      ungroup() %>% 
      group_by(Global_Name) %>% 
      mutate(prop_N_applied = ifelse(is.na(prop_N_applied), mean(prop_N_applied, na.rm = TRUE), prop_N_applied)) %>% 
      ungroup() %>% 
      dplyr::select(x,y, iso3c, prop_N_applied)
    
    
    ## applied to crops
    applied_to_crops <- left_join(N_left_gf, un_dev) %>% 
      mutate(prop_crop = case_when(Developed_Developing == "Developed" ~ 0.5,
                                   Developed_Developing == "Developing" ~ 0.95),
             prop_N_applied_crops = prop_N_applied*prop_crop) %>% 
      dplyr::select(x,y, iso3c, prop_N_applied_crops) #%>%
    #   rasterFromXYZ(crs=food_crs)
    # plot(applied_to_crops)
      
    ## save the csvs for applied manure going to crops
    write_csv(applied_to_crops, file.path(prep, paste0("animal_farm_int/STEP4_N_prop_manure_applied/applied_to_crops/",
                                 animal_name, "_",
                                 system_name,"_",
                                 product_name,
                                 "_n_prop_applied_crops.csv", sep = "")))
    
    ## applied to pastures
      applied_to_pastures <- left_join(N_left_gf, un_dev) %>% 
      mutate(prop_pasture = case_when(Developed_Developing == "Developed" ~ 0.5,
                                   Developed_Developing == "Developing" ~ 0.05),
             prop_N_applied_pastures = prop_N_applied*prop_pasture) %>% 
      select(x,y, iso3c, prop_N_applied_pastures)
      
    ## save the csvs for applied manure going to pastures
    write_csv(applied_to_pastures, file.path(prep, paste0("animal_farm_int/STEP4_N_prop_manure_applied/applied_to_pastures/",
                                 animal_name, "_",
                                 system_name,"_",
                                 product_name,
                                 "_n_prop_applied_pastures.csv", sep = "")))
    
    
}
```


Check to see if anything weird happened
```{r}
df_list <- list.files(file.path(prep,"animal_farm_int/STEP4_N_prop_manure_applied/applied_to_pastures/"), full.names = TRUE)

df <- df_list[7]

test <- read_csv(df) %>%
  dplyr::select(x,y,prop_N_applied_pastures) %>%
  rasterFromXYZ(crs = food_crs)
plot(test)

df_check <- data.frame(iso3c = NA, animal = NA, system = NA, product = NA, prop_N_applied_pastures = NA)
  
for(df in df_list){
  
#  df <- df_list[23]
    ## grab the names of all the parts from the file
    animal_name <- str_split(str_split(df, pattern = "//")[[1]][2], pattern = "_")[[1]][1]
    system_name <- str_split(str_split(df, pattern = "//")[[1]][2], pattern = "_")[[1]][2]
    product_name <- str_split(str_split(df, pattern = "//")[[1]][2], pattern = "_")[[1]][3]
     
  
  ## values are the same throughout the country so a median or mean function won't make a difference
    country_sum <- read_csv(df) %>% 
      group_by(iso3c) %>%
      mutate(animal = animal_name,
             system = system_name,
             product = product_name) %>% 
       select(iso3c, prop_N_applied_pastures, animal, system, product)
  
    df_check <- rbind(df_check, country_sum) %>% 
      filter(!is.na(iso3c)) 
    
}

## look at it

ggplot(df_check, aes(x = animal, y = prop_N_applied_pastures, color = system) ) +
  geom_point() +
  labs(title = "Proportion applied to soils")


```
