==> globalprep/prs_land-based_nutrient/v2022/STEP2_manure/STEP6_prop_manure_leaching_pasture.Rmd <==
---
title: "OHI 2022: STEP6: Prop for manure left on pastures, and prop leaching and volatized"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '../../../workflow/templates/ohi_hdr.html'
pdf_document:
  toc: true
editor_options: 
  chunk_output_type: console
---

## Summary
Creates rasters for 1) the proportion manure that leaches when left on pastures 2) the proportion manure that volatizes when left on pastures

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(here)
library(countrycode)
library(stringr)
library(vroom)
library(doParallel)

nutrient_d2021_anx <- file.path("/home/shares/ohi/git-annex/globalprep/_raw_data/FAOSTAT/manure_nutrient/d2021/")
git_manure <- file.path(here(), "globalprep/prs_land-based_nutrient/v2021/STEP2_manure")
location_tifs_path <- file.path("/home/shares/ohi/git-annex/globalprep/_raw_data/halpern_food_systems/d2021/animal_farm/location_tifs")
prep <- file.path("/home/shares/ohi/git-annex/globalprep/prs_land-based_nutrient/v2021")
git_crop <- file.path(here(), "globalprep/prs_land-based_nutrient/v2021/STEP1_crop")

source(here(git_crop, "R/common_spatial.R")) # read in necessary spatial files 

un <- read_csv(here("globalprep/spatial/v2021/UNSD_Methodology.csv")) %>% 
  dplyr::select(iso3c, Global_Name, Region_Name, Sub_region_Name, Intermediate_Region_Name)


gleam_left_pasture <- read_csv(here(git_manure, "raw/compiled_gleam_tables.csv")) %>%
  filter(mms_specific == "Pasture/Range/Paddock")

exc_df_list <- list.files(file.path(prep, "animal_farm_int/N_from_excretion_manure"),  full = TRUE)

```

## Prop left on pastures
```{r}

exc_df_distinct <- as.data.frame(exc_df_list) %>%
  rename("filepath" = "exc_df_list") %>%
  mutate(rast_file = gsub('.*/', '', filepath)) %>%
  separate(col = rast_file, sep = "_", into = c("animal_name", "system_name", "product_name", "nutrient", "process", "year")) %>%
  dplyr::select(-nutrient, -process, -year, -filepath) %>%
  distinct() 

registerDoParallel(10)   

foreach(file = 1:nrow(exc_df_distinct)) %dopar% {

     # df <- exc_df_distinct[1, ]
  
  df <- exc_df_distinct[file, ]
  
    ## grab the names of all the parts from the file
    animal_name <- df$animal_name
    system_name <- df$system_name
    product_name <- df$product_name
     
    ## This describes the proportion of waste going to each manure fate based on 
    ## country and animal system and the fraction that leaches
    mms_allocate <- gleam_left_pasture %>% 
      filter(animal == animal_name) %>% 
      filter(product == product_name) %>% 
      filter(production_system == system_name) %>% 
      dplyr::select(-ef_direct, -ef_dir_unit, -frac_gas, -percentage_leach) %>%
      unique()
    
   #length(unique(mms_allocate$iso3c)) ## missing some countries so will need to gapfill. probably later down the line when we have summarized by total N lost for each country (not mms) and then calculate a rate instead of gapfilling percentages
   
    N_left <- left_join(food_rgns_xy, mms_allocate) %>% 
      mutate(N_pasture_prop = to_mms_percentage/100) %>% 
      dplyr::select(x,y, iso3c, N_pasture_prop)
    
    write_csv(N_left,
                file.path(paste0(prep, "/animal_farm_int/STEP6_N_prop_to_pasture/",
                                 animal_name, "_",
                                 system_name,"_",
                                 product_name,
                                 "_n_prop_to_pasture.csv", sep = "")))
    
}
```

Volatization (NH3)

```{r}
bouw_table_5_2002 <- vroom::vroom(here::here(git_crop, "raw/bouwman_2002_tbl5.csv"))

volatize <- left_join(food_rgns_xy, bouw_table_5_2002) %>% 
  dplyr::select(x,y, volatize = prop_nh3_n_grassland) %>%
      arrange(x,y) %>%
  mutate(cell_id = 1:9331200) # create unique cell_id to join by 

```


Denitrification: N2O, N2
```{r}
denitr <- food_rgns_tif
denitr[denitr > 0] <- 9.125*0.008

denitr <- raster_df(denitr) %>% 
  right_join(food_rgns_xy) %>% 
  rename(denitrification = food_rgns) %>%
      arrange(x,y) %>%
  mutate(cell_id = 1:9331200) # create unique cell_id to join by 

unique(denitr$denitrification)
```

Withdraw: *0.6 of inputs after NH3 volatization

## Prop leaching
Calculate how much N is left to leach
```{r}

files <- list.files(file.path(prep, "animal_farm_int/STEP6_N_prop_to_pasture/"), full.names = TRUE)

foreach(file = files) %dopar% {

  ############  
  ## Get the mapped head counts for each system.
  ## (only really used in step 1, to check data)
  
     # file <- files[21]
  
    ## grab the names of all the parts from the file
    animal_name <- str_split(str_split(file, pattern = "//")[[1]][2], pattern = "_")[[1]][1]
    system_name <- str_split(str_split(file, pattern = "//")[[1]][2], pattern = "_")[[1]][2]
    product_name <- str_split(str_split(file, pattern = "//")[[1]][2], pattern = "_")[[1]][3]
    
    N_pasture <- read_csv(file, col_types= c("ddcn")) %>% 
          arrange(x,y) %>%
  mutate(cell_id = 1:9331200) %>% # create unique cell_id to join by 
      left_join(volatize, by = "cell_id") %>% 
      left_join(denitr, by = "cell_id") %>% 
        dplyr::select(x = x.x, y = y.x, iso3c = iso3c.x, volatize, denitrification, N_pasture_prop) %>%
      mutate(withdraw = (1 - volatize)*0.6,
             leaching = (1 - (withdraw+volatize+denitrification)),
             leaching = ifelse(leaching <0, 0, leaching),
             prop_leaching = leaching*N_pasture_prop)
    
    # test <- N_pasture %>%
    #   dplyr::select(x,y,prop_leaching) %>%
    #   rasterFromXYZ(crs = food_crs)
    # plot(test)
  
    prop_leach_gf <- left_join(N_pasture, un, by = "iso3c") %>% 
      group_by(Intermediate_Region_Name) %>% 
      mutate(prop_leaching = ifelse(is.na(prop_leaching), mean(prop_leaching, na.rm = TRUE), prop_leaching)) %>% 
      ungroup()%>% 
      group_by(Sub_region_Name) %>% 
      mutate(prop_leaching = ifelse(is.na(prop_leaching), mean(prop_leaching, na.rm = TRUE), prop_leaching)) %>% 
      ungroup() %>% 
      group_by(Region_Name) %>% 
      mutate(prop_leaching = ifelse(is.na(prop_leaching), mean(prop_leaching, na.rm = TRUE), prop_leaching)) %>% 
      ungroup() %>% 
      group_by(Global_Name) %>% 
      mutate(prop_leaching = ifelse(is.na(prop_leaching), mean(prop_leaching, na.rm = TRUE), prop_leaching)) %>% 
      ungroup() %>% 
      dplyr::select(x,y, iso3c, prop_leaching)
    
    
        write_csv(prop_leach_gf,
                file.path(paste0(prep, "/animal_farm_int/STEP6.5_N_leaching_prop_pasture/",
                                 animal_name, "_",
                                 system_name,"_",
                                 product_name,
                                 "_n_prop_leached_pasture.csv", sep = "")))
    
    
    
}

```


Check to see if anything weird happened
```{r}
df_list <- list.files(file.path(prep,"animal_farm/nutrient/STEP6.5_N_leaching_prop_pasture/"), full.names = TRUE)

df_check <- data.frame(iso3c = NA, animal = NA, system = NA, product = NA, prop_leached_pasture= NA)
  
for(df in df_list){
  
#  df <- df_list[23]
    ## grab the names of all the parts from the file
    animal_name <- str_split(str_split(df, pattern = "//")[[1]][3], pattern = "_")[[1]][1]
    system_name <- str_split(str_split(df, pattern = "//")[[1]][3], pattern = "_")[[1]][2]
    product_name <- str_split(str_split(df, pattern = "//")[[1]][3], pattern = "_")[[1]][3]
     
  ## values are the same throughout the country so a median or mean function won't make a difference
    country_sum <- read_csv(df, col_types = "ddcn") %>% 
      group_by(iso3c) %>%
      dplyr::summarise(prop_leached_pasture = median(prop_leaching, na.rm = TRUE)) %>% 
      mutate(animal = animal_name,
             system = system_name,
             product = product_name) %>% 
       select(iso3c, prop_leached_pasture , animal, system, product)
  
    df_check <- rbind(df_check, country_sum) %>% 
      filter(!is.na(iso3c)) 
    
}

## look at it

ggplot(df_check, aes(x = animal, y = prop_leached_pasture, color = system) ) +
  geom_point() +
  labs(title = "Proportion of all manure leached from left on pasture")


```


Let's look at just the USA cows no feedlot

```{r}
df_list <- list.files(file.path(prep, "animal_farm_int/STEP6.5_N_leaching_prop_pasture/"), full.names = TRUE)
df_list <- str_subset(pattern = "cows", df_list)
df_file <- df_list[4]

    ## grab the names of all the parts from the file
    category_name <- str_split(df_file, pattern = "//")[[1]][2]
  
  ## values are the same throughout the country so a median or mean function won't make a difference
    country_sum_mixed <- read_csv(df_file, col_types = "ddcn") %>% 
      group_by(iso3c) %>% 
      dplyr::summarise(prop_pasture_leaching = mean(prop_leaching, na.rm = TRUE)) %>% 
      mutate(category = category_name,
             animal_system_product = word(category, 1,3, sep="_")) %>% 
       dplyr::select(iso3c,prop_pasture_leaching, animal_system_product) %>% 
      filter(iso3c == "USA")
    
    df_file <- df_list[2]

    ## grab the names of all the parts from the file
    category_name <- str_split(df_file, pattern = "//")[[1]][2]
  
  ## values are the same throughout the country so a median or mean function won't make a difference
    country_sum_pasture <- read_csv(df_file, col_types = "ddcn") %>% 
      group_by(iso3c) %>% 
      dplyr::summarise(prop_pasture_leaching = median(prop_leaching, na.rm = TRUE)) %>% 
      mutate(category = category_name,
             animal_system_product = word(category, 1,3, sep="_")) %>% 
       dplyr::select(iso3c,prop_pasture_leaching, animal_system_product) %>% 
      filter(iso3c == "USA")
    
  country_sum <- rbind(country_sum_pasture, country_sum_mixed) 

write_csv(country_sum, here::here(git_manure, "data_check/usa_cows_leached_pasture.csv"))
```


## Prop volatized
Calculate how much N volatizes
```{r}

files <- list.files(file.path(prep, "animal_farm_int/STEP6_N_prop_to_pasture/"), full.names = TRUE)

registerDoParallel(5)

foreach(file = files) %dopar% {

  ############  
  ## Get the mapped head counts for each system.
  ## (only really used in step 1, to check data)
  
     # file <- files[20]
  
    ## grab the names of all the parts from the file
    animal_name <- str_split(str_split(file, pattern = "//")[[1]][2], pattern = "_")[[1]][1]
    system_name <- str_split(str_split(file, pattern = "//")[[1]][2], pattern = "_")[[1]][2]
    product_name <- str_split(str_split(file, pattern = "//")[[1]][2], pattern = "_")[[1]][3]
    
    N_pasture <- read_csv(file, col_types= c("ddcn")) %>% 
                arrange(x,y) %>%
  mutate(cell_id = 1:9331200) %>% # create unique cell_id to join by 
      left_join(volatize, by = "cell_id") %>% 
      dplyr::select(x = x.x, y = y.x, iso3c, volatize, N_pasture_prop) %>%
      mutate(prop_volt = volatize*N_pasture_prop) 
  
    prop_volt_gf <- left_join(N_pasture, un, by = "iso3c") %>% 
      group_by(Intermediate_Region_Name) %>% 
      mutate(prop_volt = ifelse(is.na(prop_volt), mean(prop_volt, na.rm = TRUE), prop_volt)) %>% 
      ungroup()%>% 
      group_by(Sub_region_Name) %>% 
      mutate(prop_volt = ifelse(is.na(prop_volt), mean(prop_volt, na.rm = TRUE), prop_volt)) %>% 
      ungroup() %>% 
      group_by(Region_Name) %>% 
      mutate(prop_volt = ifelse(is.na(prop_volt), mean(prop_volt, na.rm = TRUE), prop_volt)) %>% 
      ungroup() %>% 
      group_by(Global_Name) %>% 
      mutate(prop_volt = ifelse(is.na(prop_volt), mean(prop_volt, na.rm = TRUE), prop_volt)) %>% 
      ungroup() %>% 
      dplyr::select(x,y, iso3c, prop_volt)
    
    
        write_csv(prop_volt_gf,
                file.path(paste0(prep, "/animal_farm_int/STEP6.5_N_volatizing_prop_pasture/",
                                 animal_name, "_",
                                 system_name,"_",
                                 product_name,
                                 "_n_prop_volatized_pasture.csv", sep = "")))
    
    
    
}

```
